+++ FrontEnd/src/views/Security/CameraTwo.jsx
@@ -1,5 +1,6 @@
 import React, { Component } from "react";
 import { Grid, Col, Row } from "react-bootstrap";
+import Button from 'components/CustomButton/CustomButton.jsx';
 // react component used to create charts
 import ChartistGraph from "react-chartist";
 // react components used to create a SVG / Vector map
@@ -21,64 +22,21 @@ import {
 } from "variables/Variables.jsx";
 
 class CameraTwo extends Component {
-  createTableData() {
-    var tableRows = [];
-    for (var i = 0; i < table_data.length; i++) {
-      tableRows.push(
-        <tr key={i}>
-          <td>
-            <div className="flag">
-              <img src={table_data[i].flag} alt="us_flag" />
-            </div>
-          </td>
-          <td>{table_data[i].country}</td>
-          <td className="text-right">{table_data[i].count}</td>
-          <td className="text-right">{table_data[i].percentage}</td>
-        </tr>
-      );
-    }
-    return tableRows;
-  }
   render() {
     return (
-      <div className="main-content">
+      <div className="main-content camera-container">
         <Grid fluid>
           <Row>
             <Col lg={3} sm={6}>
-              <StatsCard
-                bigIcon={<i className="pe-7s-server text-warning" />}
-                statsText="Capacity"
-                statsValue="105GB"
-                statsIcon={<i className="fa fa-refresh" />}
-                statsIconText="Updated now"
-              />
+              <video src="http://192.168.8.223:8000" />
+            </Col>
+          </Row>
+          <Row>
+            <Col lg={3} sm={6}>
+              <Button bsStyle="danger">Stop Attack</Button>
             </Col>
             <Col lg={3} sm={6}>
-              <StatsCard
-                bigIcon={<i className="pe-7s-wallet text-success" />}
-                statsText="Revenue"
-                statsValue="$1,345"
-                statsIcon={<i className="fa fa-calendar-o" />}
-                statsIconText="Last day"
-              />
-            </Col>
-            <Col lg={3} sm={6}>
-              <StatsCard
-                bigIcon={<i className="pe-7s-graph1 text-danger" />}
-                statsText="Errors"
-                statsValue="23"
-                statsIcon={<i className="fa fa-clock-o" />}
-                statsIconText="In the last hour"
-              />
-            </Col>
-            <Col lg={3} sm={6}>
-              <StatsCard
-                bigIcon={<i className="fa fa-twitter text-info" />}
-                statsText="Followers"
-                statsValue="+45"
-                statsIcon={<i className="fa fa-refresh" />}
-                statsIconText="Updated now"
-              />
+              <Button bsStyle="default">False Alarm</Button>
             </Col>
           </Row>
         </Grid>